Seafile: add deprecation notice & rename json file
Current Seafile rockon works only for seafile 6.x because starting with
v7.0, seafile have adjusted seafile-docker image to use multiple
containers. But to prevent accidental old-fashioned upgrades, the new
docker image has a new name called seafile-mc (Seafile Multiple
Container). Thus it is safe to continue using this rockon if someone
wishes to stay on Seafile v6.3.4.

However, version 6 is no longer maintained so we add a deprecation
notice and a link on how to migrate. Moreover, we rename the json
file to disambiguate with future version v7+.