Move scrutiny to official image, altering options accordingly #329
Move from linuxserver.io to official master-omnibus image and remove
redundant settings.

## Includes:
- Rewording and simplifying instructions.
- Capitalisation fixes.
- Removal of SCRUTINY_API_ENDPOINT option (to simplify).
- Adding COLLECTOR_CRON_SCHEDULE set option. It is often desired to
control of when a collection is enacted. This also surfaces that this is
only done once per day.
- Update Rock-on json file version number.
- Correct prior mix of tab and spaces.

### Caveats:
The "--privileged" option was found to be required but was not
referenced in the upstream documentation! We already have the two
referenced "--cap-add".