add bareos-webui-php-fpm-to-director docker network #94
It appears that we need this network in order that the PHP
code this container carries can also resolve our director.
We have a current limitation regarding unique `source_container`
network config entries, which prevents us from simply having
`bareos-webui-php-fpm` as an additional source_container entry
in the "bareos-director" network. This was encountered when adding
the bareos-webui network.