Update paperless-ngx.json
I made a mistake in the volume part for paperless. I have fixed this. Should have no impact on usage and update as only the export and import shares were changed. Added PAPERLESS_SECRET_KEY, PAPERLESS_URL and PAPERLESS_CSRF_TRUSTED_ORIGINS which can all be left empty but are useful if a nginx or similar is in use.