Remove all but one password re-enter requirements #94
As a server-set, with ordered install, we can inherit
(via the shared config arrangement) all intra Bareos
service passwords. File and Storage containers are now
able to self-config when uses in this intended AIO setup.
This is akin to the rpms default behaviour of reading
each-others config in order to auto-setup the initial server
set.

Includes minor Description rewording.